requirement.txt
pytorch
numpy
matplotlib
Pillow
scipy
skimage
imgaug